move combo editor, refactor context
