+++ libs/sr/page-team/src/context/TeamContext.ts
@@ -1,16 +1,14 @@
-import type { Team, TeammateDatum } from '@genshin-optimizer/sr/db'
+import type { Team } from '@genshin-optimizer/sr/db'
 import { createContext, useContext } from 'react'
 
 export type TeamContextObj = {
   teamId: string
   team: Team
-  teammateDatum: TeammateDatum
 }
 
 export const TeamContext = createContext({
   teamId: '',
   team: {},
-  teammateDatum: { buildType: 'equipped', buildId: '' },
 } as TeamContextObj)
 
 export function useTeamContext() {
